% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_drules.R
\name{apply_drules}
\alias{apply_drules}
\title{Run commands dynamically trough mutate}
\usage{
apply_drules(df, column, cmd, returnDF = F)
}
\arguments{
\item{df}{A data frame}

\item{column}{The selected column for mutate}

\item{cmd}{The command to be run as a string}

\item{returnDF}{If the full data frame is to be returned or just a vector}
}
\value{
A vector with the results or the full DF with the new column
}
\description{
Create or modify a column with mutate and a command passed as a string.
}
\examples{
df <- apply_drules(df, "mycol", "case_when(other_col < 2 ~ 'low', TRUE ~ 'high')", returnDF = TRUE);
}
