% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsqts-package.R
\docType{package}
\name{bsqts-package}
\alias{bsqts}
\alias{bsqts-package}
\title{bsqts: bsqts}
\description{
A handful of recurrent functions for biological data science
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pbousquets/bsqts}
  \item Report bugs at \url{https://github.com/pbousquets/bsqts/issues}
}

}
\author{
\strong{Maintainer}: Pablo Bousquets \email{pbousquetsm@gmail.com} (\href{https://orcid.org/0000-0002-2969-008X}{ORCID}) [copyright holder]

}
\keyword{internal}
